+++ src/app/components/Common/MainLayout.js
@@ -1,6 +1,8 @@
 import React from 'react';
-import { flexContainer } from '../../styles';
-  
+
+import Menu from '../Menu';
+import { flexContainer } from '../../styles/style.js';
+
 const content = {
   width: 750
 };
@@ -9,17 +11,18 @@ const side = {
   flex: 1
 };
 
-export default MainLayout = ({mainComponent}) => {
+const MainLayout = ({mainComponent}) => {
   return (
     <div style={flexContainer}>
       <div style={side}>
         <Menu isMenuOpen={true}/>
       </div>
       <div style={content}>
-        {{ MainComponent }}
+        {mainComponent}
       </div>
       <div style={side}></div>
     </div>
   );
 }
 
+export default MainLayout;
