+++ src/app/components/Common/NovelList/NovelCard/ReadButton.js
@@ -3,10 +3,12 @@ import React from 'react';
 import IconButton from 'material-ui/IconButton';
 import ActionDone from 'material-ui/svg-icons/action/done';
 
-export default ReadButton = ({readCard}) => {
+const ReadButton = ({readCard}) => {
   return (
     <IconButton onClick={readCard}>
       <ActionDone />
     </IconButton>
   );
 };
+
+export default ReadButton;
