+++ src/app/components/Novels/Novels.js
@@ -4,7 +4,7 @@ import { connect } from 'react-redux';
 import MainLayout from '../Common/MainLayout';
 
 import SearchBar from './SearchBar';
-import NovelList from './NovelList';
+import NovelList from '../Common/NovelList';
 
 
 import { initializeDbListWith } from '../../actions/listActions';
@@ -17,20 +17,12 @@ class Novels extends React.Component {
       .then((v)=> this.props.initializeDbListWith(v));
   }
   render () {
-    const content = {
-      width: 750
-    };
-
-    const side = {
-      flex: 1
-    };
-
     return (
       <div>
         <header>
           <SearchBar/>
         </header>
-        <MainLayout mainComponent={NovelList}/>
+        <MainLayout mainComponent={(<NovelList/>)}/>
       </div>
     );
   }
