+++ src/app/components/Novels/SearchBar/SearchButton.js
@@ -4,13 +4,15 @@ import IconButton from 'material-ui/IconButton';
 import ActionSearch from 'material-ui/svg-icons/action/search';
 import { white } from 'material-ui/styles/colors';
 
-import { pointer } from '../../../styles';
+import { pointer } from '../../../styles/style.js';
 
 
-export default SearchButton = ({searchMethod}) => {
+const SearchButton = ({searchMethod}) => {
   return (
     <IconButton onClick={searchMethod} tooltip='Search'>
       <ActionSearch color={ white } style={pointer}　/>
     </IconButton>
   )
 };
+
+export default SearchButton;
