--- src/app/components/Novels/SearchForm/index.js
@@ -1,3 +1,0 @@
-import SearchForm from './SearchForm';
-
-export default SearchForm;
