+++ src/index.js
@@ -6,9 +6,9 @@ import { Routes, store } from './config'
 
 import injectTapEventPlugin from 'react-tap-event-plugin';
 import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
-import theme from './styles/theme';
+import theme from './app/styles/theme';
 
-import './styles/index.css';
+import './app/styles/index.css';
 
 // Needed for onTouchTap
 // http://stackoverflow.com/a/34015469/988941
